MovieRentalAPI/requirements.txt
fastapi
python-dotenv
uvicorn
pydantic
pyjwt